go/common/sgx: Move encave attributes type to the sgx package
This was only used for IAS, but now that SIGSTRUCT generation is a
thing, it's better in the common sgx package.
