+++ go/oasis-node/cmd/debug/byzantine/steps.go
@@ -77,8 +77,8 @@ func initFakeCapabilitiesSGX() (signature.Signer, *node.Capabilities, error) {
 			Version: 1,
 		},
 		Report: ias.Report{
-			Attributes: ias.Attributes{
-				Flags: ias.AttributeDebug,
+			Attributes: sgx.Attributes{
+				Flags: sgx.AttributeDebug,
 			},
 			MRENCLAVE: enclaveIdentity.MrEnclave,
 			MRSIGNER:  enclaveIdentity.MrSigner,
